feat: Cart Fixing Detail

+++ controllers/CartControllers.js
@@ -59,10 +59,14 @@ class CartController {
       if (!menuItem) {
         return res.status(404).json({ message: "Menu item not found" });
       }
+
+      // Check if the menu item is available
+      if(quantity <= 0){
+        return res.status(400).json({ message: "Quantity must be greater than 0" });
+      }
+
       if (menuItem.available_stock < quantity) {
-        return res
-          .status(400)
-          .json({ message: "Not enough stock available" });
+        return res.status(400).json({ message: "Not enough stock available" });
       }
       // Check if the item already exists in the cart
       const existingOrderItem = await Order_Item.findOne({
@@ -72,13 +76,20 @@ class CartController {
         // If it exists, update the quantity
         existingOrderItem.quantity += quantity;
         await existingOrderItem.save();
-      
+
         // Update total_price as well
         cart.total_price += quantity * menuItem.price;
         await cart.save();
-      
+
+        menuItem.available_stock -= quantity;
+        await menuItem.save();
+
         return res.status(200).json(existingOrderItem);
       }
+
+      menuItem.available_stock -= quantity; // Reduce the available stock
+      await menuItem.save(); // Save the updated menu item
+
       // If it doesn't exist, create a new order item
       const orderItem = await Order_Item.create({
         order_id: cart.id,
@@ -118,7 +129,23 @@ class CartController {
         return res.status(404).json({ message: "Cart not found" });
       }
 
-      const orderItem = await Order_Item.destroy({
+      // reduce the total price of the cart
+      const orderItem = await Order_Item.findOne({
+        where: { id: order_item_id, order_id: cart.id },
+      });
+      if (!orderItem) {
+        return res.status(404).json({ message: "Order item not found" });
+      }
+
+      const menuItem = await Menu_Item.findByPk(orderItem.menu_item_id);
+      menuItem.available_stock += orderItem.quantity; // Increase the available stock
+      await menuItem.save(); // Save the updated menu item
+
+      cart.total_price -= orderItem.quantity * orderItem.price_per_item;
+      await cart.save();
+
+      // remove the order item from the cart
+      await Order_Item.destroy({
         where: { id: order_item_id, order_id: cart.id },
       });
 
@@ -129,6 +156,65 @@ class CartController {
       return res.status(200).json({ message: "Order item removed from cart" });
     } catch (error) {
       console.error("Error removing from cart:", error);
+      return res.status(500).json({ message: "Internal server error" });
+    }
+  }
+
+  static async updateOrderItemQuantity(req, res) {
+    try {
+      const userId = req.user.id; // Assuming you have user ID in the request
+      const { order_item_id } = req.params;
+      const { quantity } = req.body;
+
+      if(quantity <= 0){
+        return res.status(400).json({ message: "Quantity must be greater than 0" });
+      }
+
+      if (!order_item_id || !quantity) {
+        return res
+          .status(400)
+          .json({ message: "Order item ID and quantity are required" });
+      }
+
+      const cart = await Order.findOne({
+        where: { user_id: userId, status: "cart" },
+      });
+
+      if (!cart) {
+        return res.status(404).json({ message: "Cart not found" });
+      }
+
+      const orderItem = await Order_Item.findOne({
+        where: { id: order_item_id, order_id: cart.id },
+      });
+
+      if (!orderItem) {
+        return res.status(404).json({ message: "Order item not found" });
+      }
+
+      // Update the quantity and total price
+      const menuItem = await Menu_Item.findByPk(orderItem.menu_item_id);
+      if (menuItem.available_stock < quantity) {
+        return res.status(400).json({ message: "Not enough stock available" });
+      }
+      
+      const gapQuantity = quantity - orderItem.quantity;
+      console.log("Gap Quantity:", gapQuantity);
+      // Update stock
+      if (gapQuantity !== 0) {
+        menuItem.available_stock -= gapQuantity;
+        await menuItem.save();
+      }
+
+      cart.total_price += (quantity - orderItem.quantity) * menuItem.price;
+      await cart.save();
+
+      orderItem.quantity = quantity;
+      await orderItem.save();
+
+      return res.status(200).json(orderItem);
+    } catch (error) {
+      console.error("Error updating order item quantity:", error);
       return res.status(500).json({ message: "Internal server error" });
     }
   }
